+++ citta_console/__init__.py
@@ -26,4 +26,4 @@ __all__ = [
     "render_dashboard",
 ]
 
-__version__ = "0.7.0"
+__version__ = "0.8.0"